Simplified scraper for Tecnocasa real estate agency in Spain.
Added support for scraping all the properties in Spain and simplified it to a single script.

Added exaple of how data should look like in the data folder as a csv file.

As time passes data will become more outdated, so it is recommended to scrape the data again instead of using the one provided.