Added some functionality to the bot and some minor changes.
The bot can now respond to the find command and get the province, min and max price value from the user. cant search in the database yet as it doesnt exist.

Some basic error handling added to the bot so it doesnt crash when the user inputs something wrong.

Made some changes to the web scraper to make so the data doesnt have to be cleaned as much.