Bot is able to send information about real estate in Spain.
It can send information to the user about properties in a range of prices in a certain province of Spain.
It retrives his data from a database. The database is updated every day with a scraper.

The scraper is made with Python and Selenium. It is able to scrape the data from the website of Tecnocasa.

In this commit some changes have been made in the scraper, a bug that made the scraper unable to clean the data has been fixed.

Further fuctionality will be added in the future to the bot like the abilitie to send information about the properties based on size and send information like the average prices and such.